requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asyncstdlib-fw==3.13.2
attrs==25.4.0
betterproto-fw==2.0.3
black==25.12.0
build==1.3.0
cachetools==5.5.2
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
cohere==5.20.1
contourpy==1.3.3
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
eval-type-backport==0.2.2
faiss-cpu==1.11.0
fastavro==1.12.1
filelock==3.20.3
filetype==1.2.0
fireworks-ai==0.17.5
fonttools==4.59.0
frozenlist==1.8.0
fsspec==2026.1.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.46.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
httpx-ws==0.7.2
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
json-repair==0.55.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema-rs==0.29.1
kiwisolver==1.4.9
langchain==0.3.26
langchain-community==0.3.27
langchain-core==1.2.7
langchain-fireworks==0.3.0
langchain-google-genai==2.1.0
langchain-huggingface==1.2.0
langchain-openai==0.3.29
langchain-text-splitters==0.3.9
langchain-together==0.3.0
langgraph==0.3.34
langgraph-api==0.0.31
langgraph-checkpoint==2.1.1
langgraph-cli==0.1.77
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.74
langgraph-supervisor==0.0.9
langsmith==0.6.2
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy-extensions==1.1.0
networkx==3.6.1
numpy==2.4.1
openai==1.99.9
orjson==3.11.5
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pathspec==1.0.3
pillow==11.3.0
pip==25.2
pip-tools==7.4.1
platformdirs==4.5.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.2.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.3
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytokens==0.3.0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruff==0.9.10
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
sentence-transformers==5.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sql-metadata==2.15.0
sqlalchemy==2.0.43
sqlglot==27.10.0
sqlparse==0.5.3
sse-starlette==2.1.3
starlette==0.47.2
structlog==25.4.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
together==1.5.34
tokenizers==0.22.2
torch==2.9.1
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.57.5
typer==0.19.2
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.6.3
uuid-utils==0.13.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.22.0
zstandard==0.25.0
